pyproject.toml
[tool.poetry]
name = "backups"
version = "4.3.0"
description = "A simple productivity platform"
authors = ["Dimitrios Strantsalis <dstrants@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3"
djangorestframework = "^3.12.1"
django-extensions = "^3.0.9"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
requests = "^2.24.0"
requests-oauthlib = "^1.3.0"
sentry_sdk = "^1.0.0"
social-auth-app-django = "^4.0.0"
apache-libcloud = "^3.2.0"
google-api-python-client = "^1.12.5"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
PyGithub = "^1.53"
rich = "^9.1.0"
pendulum = "^2.1.2"
pydantic = "^1.7.1"
python-dotenv = "^0.15.0"
django-redis = "^4.12.1"
todoist-python = "^8.1.3"
django-dramatiq = "^0.9.1"
periodiq = "^0.12.1"
django-periodiq = {git = "https://github.com/Sovetnikov/django_periodiq"}
django-anymail = {version = "^8.2", extras = ["mailgun"]}
Pillow = "^8.1.2"
django-dbbackup = "^3.3.0"
dropbox = "^11.5.0"
django-storages = "^1.11.1"
PyYAML = "^5.4.1"
slack-sdk = "^3.7.0"
boto3 = "^1.18.0"
Faker = "^8.12.0"
Markdown = "^3.3.4"
pymongo = "^3.12.1"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]
pylama = "^7.7.1"
pytest = "^6.1.2"
pytest-django = "^4.1.0"
coverage = "^5.5"
factory-boy = "^3.2.0"
faker-microservice = "^2.0.0"
isort = "^5.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
